requirements.txt
 fastapi
uvicorn[standard]
python-dotenv
google-cloud-storage
google-generativeai
tenacity
pandas
openpyxl
pytesseract
Pillow
pdf2image
python-multipart
